backend/requirements.txt
Flask
Flask-JWT-Extended
pymongo
python-dotenv
requests
werkzeug
